+++ sanity/schemas/documents/page.ts
@@ -30,6 +30,7 @@ export default defineType({
 		select: {
 			title: 'title',
 			slug: 'metadata.slug.current',
+			media: 'metadata.image',
 		},
 		prepare: ({ title, slug }) => ({
 			title,